Single Post Page
Should work for displaying the user and individual post, still working
on adding the vote and comment buttons, as well as having them all
display. Unsure about what the smbdeleteAAA thing is all about… i
unchecked it but it won’t go away. I also undid the little bit of
modifying I had started on the header, as I didn’t think it was needed
to alter the logout button